rag/requirements.txt
# Framework Web et Serveur
fastapi==0.111.0
uvicorn[standard]==0.27.1
gunicorn==23.0.0

# Base de données et Cache
redis==5.0.1
motor>=3.0.0,<4.0.0
pymongo==4.6.3

# IA et Traitement des données
openai==1.67.0
numpy<2.0.0

# Utilitaires et Sécurité
pydantic>=1.10.0,<3.0.0
python-dotenv>=1.0.0
aiohttp==3.10.11
python-jose[cryptography]==3.4.0
python-multipart==0.0.18
slowapi==0.1.8
bech32==1.2.0

# Tests
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# RAG
llama-index
faiss-cpu
rank-bm25
tiktoken
pypdf
langchain
langchain-community

nest_asyncio

httpx>=0.25.0
asyncio>=3.4.3
uuid>=1.30

# Nouvelles dépendances
networkx>=3.2.1
plotly>=5.18.0
pandas>=2.2.3
scipy>=1.13.1

# Visualisation
streamlit>=1.29.0
